api/requirements.txt
azure-functions
openai>=1.60.0
python-dotenv>=1.0.0
azure-cosmos>=4.9.0

